requirements.txt
colorama==0.4.6
coverage==7.10.2
iniconfig==2.1.0
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
pytest==8.4.1
pytest-cov==6.2.1
python-dotenv==1.1.1
tabulate==0.9.0
